Docker

+++ delphin_6_automation/logging/ribuild_logger.py
@@ -12,52 +12,52 @@ import platform
 
 # RiBuild Modules:
 
-try:
-    from delphin_6_automation.database_interactions.auth import gmail
-except ModuleNotFoundError:
-    gmail = {'mail': None, 'password': None}
-
 # -------------------------------------------------------------------------------------------------------------------- #
 # LOGGERS
 
+loggers = {}
 
-def ribuild_logger(name):
+
+def ribuild_logger(name: str='delphin_6_automation'):
+    global loggers
 
     source_folder = os.environ.get("_MEIPASS2", os.path.abspath("."))
 
-    # create logger
-    logger = logging.getLogger(name)
-    logger.setLevel(logging.DEBUG)
-    formatter = logging.Formatter('%(asctime)s - %(levelname)s - %(threadName)s - %(funcName)s - %(message)s')
+    if loggers.get(name):
+        return loggers.get(name)
 
-    # create console handler and set level to debug
-    if os.path.exists(f'{source_folder}/{name}.log'):
-        try:
-            os.remove(f'{source_folder}/{name}.log')
-        except PermissionError:
-            pass
+    else:
+        # create logger
+        logger = logging.getLogger(name)
+        logger.setLevel(logging.DEBUG)
+        formatter = logging.Formatter('%(asctime)s - %(levelname)s - %(filename)s - %(funcName)s - %(message)s')
 
-    # File Handler
-    fh = logging.FileHandler(f'{source_folder}/{name}.log')
-    fh.setLevel(logging.DEBUG)
-    fh.setFormatter(formatter)
-    logger.addHandler(fh)
+        # create console handler and set level to debug
+        if os.path.exists(f'{source_folder}/{name}.log'):
+            try:
+                os.remove(f'{source_folder}/{name}.log')
+            except PermissionError:
+                pass
 
-    # Notification Handler
-    message_dict = {'to': 'ocni@dtu.dk',
-                    'subject': f'Fatal Error on {platform.node()}',
-                    'username': gmail['mail'],
-                    'password': gmail['password']}
-    nh = NotificationHandler('gmail', defaults=message_dict)
-    nh.setLevel(logging.WARNING)
-    nh.setFormatter(formatter)
-    logger.addHandler(nh)
+        # File Handler
+        fh = logging.FileHandler(f'{source_folder}/{name}.log')
+        fh.setLevel(logging.DEBUG)
+        fh.setFormatter(formatter)
+        logger.addHandler(fh)
 
-    # Stream Handler
-    sh = logging.StreamHandler()
-    stream_formatter = logging.Formatter('%(threadName)s: %(message)s')
-    sh.setFormatter(stream_formatter)
-    sh.setLevel(logging.INFO)
-    logger.addHandler(sh)
+        # Notification Handler
+        message_dict = {'message': f'Fatal Error on {platform.node()}',
+                        'webhook_url': 'https://hooks.slack.com/services/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'}
+        nh = NotificationHandler('slack', defaults=message_dict)
+        nh.setLevel(logging.WARNING)
+        nh.setFormatter(formatter)
+        logger.addHandler(nh)
 
-    return logger
+        # Stream Handler
+        sh = logging.StreamHandler()
+        sh.setFormatter(formatter)
+        sh.setLevel(logging.DEBUG)
+        logger.addHandler(sh)
+        loggers[name] = logger
+
+        return logger
